refactor :: authLayout into components/layout
- authLayout 폴더 위치 변경
