+++ src/components/layout/authLayout/AuthStepComponentLayout/ComponentLayout.tsx
