+++ src/components/layout/authLayout/AuthStepComponentLayout/ContentLayout.tsx
