+++ src/components/layout/authLayout/AuthStepComponentLayout/StepHeader.tsx
