+++ src/navigation/stacks/MainStack.tsx
@@ -5,9 +5,9 @@ import AdminSignupScreen from "@/screens/auth/signup/AdminSignupScreen";
 import UserSignupScreen from "@/screens/auth/signup/UserSignupScreen";
 import SignupRoleSelectScreen from "@/screens/auth/SignupRoleSelectScreen";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import { MainTabParamList } from "./type";
+import { MainTabParamList } from "../type";
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import BottomBar from "./BottomTab";
+import BottomBar from "../tabs/BottomTab";
 
 const Tab = createBottomTabNavigator<MainTabParamList>();
