+++ src/screens/auth/SignupRoleSelectScreen.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/native';
-import ActionLayout from "@/components/auth/authLayout/AuthStepComponentLayout/ActionLayout";
-import StepHeader from "@/components/auth/authLayout/AuthStepComponentLayout/StepHeader";
-import AuthStepLayout from "@/components/auth/authLayout/AuthStepLayout";
+import ActionLayout from "@/components/layout/authLayout/AuthStepComponentLayout/ActionLayout";
+import StepHeader from "@/components/layout/authLayout/AuthStepComponentLayout/StepHeader";
+import AuthStepLayout from "@/components/layout/authLayout/AuthStepLayout";
 import Typography from "@/components/common/typography/Typography";
 import btn_student_default from "@/assets/btn_role-student_default.png"
 import btn_student_selected from "@/assets/btn_role-student_selected.png"
