+++ src/screens/auth/onboarding/steps/CompleteStep.tsx
@@ -1,4 +1,4 @@
-import ContentLayout from "@/components/auth/authLayout/AuthStepComponentLayout/ContentLayout";
+import ContentLayout from "@/components/layout/authLayout/AuthStepComponentLayout/ContentLayout";
 import Typography from "@/components/common/typography/Typography";
 
 export default function CompleteStep() {
