+++ src/screens/auth/password/FindPasswordScreen.tsx
@@ -1,6 +1,6 @@
-import ActionLayout from "@/components/auth/authLayout/AuthStepComponentLayout/ActionLayout"
-import StepHeader from "@/components/auth/authLayout/AuthStepComponentLayout/StepHeader"
-import AuthStepLayout from "@/components/auth/authLayout/AuthStepLayout"
+import ActionLayout from "@/components/layout/authLayout/AuthStepComponentLayout/ActionLayout"
+import StepHeader from "@/components/layout/authLayout/AuthStepComponentLayout/StepHeader"
+import AuthStepLayout from "@/components/layout/authLayout/AuthStepLayout"
 import KeyboardDismiss from "@/components/common/KeyboardDismiss"
 import { useSignupStepControl } from "@/hooks/useSignupStepControl"
 import { FindContentProps, FindForm } from "@/types/index"
