+++ src/screens/auth/password/steps/EmailStep.tsx
@@ -1,6 +1,6 @@
 import Typography from "@/components/common/typography/Typography"
 import { useEffect, useState } from "react"
-import ContentLayout from "@/components/auth/authLayout/AuthStepComponentLayout/ContentLayout";
+import ContentLayout from "@/components/layout/authLayout/AuthStepComponentLayout/ContentLayout";
 import DefaultInput from "@/components/auth/authInput/DefaultInput";
 import { isValidEmail } from "@/utils/isValidEmail";
