+++ src/screens/auth/signup/steps/IdStep.tsx
@@ -3,7 +3,7 @@ import DefaultInput from "@/components/auth/authInput/DefaultInput";
 import Typography from "@/components/common/typography/Typography";
 import { colorStyle } from "@/styles/colorStyle";
 import { useEffect, useState } from "react";
-import ContentLayout from "@/components/auth/authLayout/AuthStepComponentLayout/ContentLayout";
+import ContentLayout from "@/components/layout/authLayout/AuthStepComponentLayout/ContentLayout";
 import { Pressable } from "react-native";
 import { isValidEmail } from "@/utils/isValidEmail";
